requirements.txt
# --- Core Frameworks ---
langchain
langgraph
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
fastapi
uvicorn
streamlit
pydantic
beautifulsoup4
requests
selenium
python-dotenv
langchain-openai
openai

# --- Guardrails AI & Validators ---
guardrails-ai
# Underlying packages for hub://guardrails/detect_pii and detect_prompt_injection
guardrails-grhub-detect-pii
guardrails

# --- Validator Dependencies ---
presidio-analyzer
presidio-anonymizer